backEndApiFinanceApp/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
et_xmlfile==2.0.0
frozendict==2.4.4
html5lib==1.1
idna==3.7
lxml==5.2.2
Markdown==3.6
multitasking==0.0.11
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.2
peewee==3.17.5
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.40
